refactor(web): group routes under protected
Move protected pages into Next.js route-group structure under\nweb/src/app/(protected), keeping page behavior while isolating\nauthenticated layout boundaries for clearer app composition.\n\nRefs #173
