+++ web/src/app/(protected)/admin/instance-sizes/page.tsx
@@ -1,16 +1,16 @@
 'use client';
 
 /**
- * Template management — admin page (read-only list for V1).
+ * InstanceSize management — admin page (read-only list for V1).
  *
- * OpenAPI: GET /templates
- * master-flow Stage 3, Step 3: Template defines OS image + cloud-init.
- * ADR-0018: Templates stored in DB, define OS image source and cloud-init only.
+ * OpenAPI: GET /instance-sizes
+ * master-flow Stage 3, Step 4: InstanceSize defines hardware resources.
+ * ADR-0018: Hardware capability requirements (GPU/SR-IOV/Hugepages) in InstanceSize.
  *
  * Search best practices (Context7 / React docs / Ant Design v5):
  * - useDeferredValue: React-recommended debounce for search input
  * - filterDropdown: Ant Design column-level search with highlight
- * - filters + onFilter: Ant Design column-level enum filter for OS Family / Status
+ * - filters + onFilter: Ant Design column-level enum filter for capabilities / status
  * - useMemo: cache filtered results to avoid re-render on unchanged data
  */
 import { useState, useDeferredValue, useMemo, useRef } from 'react';
@@ -28,8 +28,9 @@ import type { InputRef } from 'antd';
 import type { ColumnsType, FilterDropdownProps } from 'antd/es/table/interface';
 import {
     ReloadOutlined,
-    FileTextOutlined,
+    HddOutlined,
     SearchOutlined,
+    ThunderboltOutlined,
 } from '@ant-design/icons';
 import { useTranslation } from 'react-i18next';
 import { useApiGet } from '@/hooks/useApiQuery';
@@ -38,18 +39,19 @@ import type { components } from '@/types/api.gen';
 
 const { Title, Text } = Typography;
 
-type Template = components['schemas']['Template'];
-type TemplateList = components['schemas']['TemplateList'];
+type InstanceSize = components['schemas']['InstanceSize'];
+type InstanceSizeList = components['schemas']['InstanceSizeList'];
 
-const OS_COLOR_MAP: Record<string, string> = {
-    linux: 'green',
-    windows: 'blue',
-    centos: 'orange',
-    ubuntu: 'geekblue',
-    debian: 'purple',
-    rhel: 'red',
-    fedora: 'cyan',
-};
+/**
+ * Format memory from MB to human-readable (GB if >= 1024).
+ */
+function formatMemory(mb: number): string {
+    if (mb >= 1024) {
+        const gb = mb / 1024;
+        return `${Number.isInteger(gb) ? gb : gb.toFixed(1)} GB`;
+    }
+    return `${mb} MB`;
+}
 
 /**
  * Highlight matching text within a string.
@@ -73,9 +75,20 @@ function highlightText(text: string, highlight: string): React.ReactNode {
     );
 }
 
-export default function TemplatesPage() {
+/**
+ * Derive capability tags for a given InstanceSize record.
+ */
+function getCapabilityLabels(record: InstanceSize): string[] {
+    const caps: string[] = [];
+    if (record.requires_gpu) caps.push('GPU');
+    if (record.requires_sriov) caps.push('SR-IOV');
+    if (record.requires_hugepages) caps.push('Hugepages');
+    if (record.dedicated_cpu) caps.push('Dedicated CPU');
+    return caps;
+}
+
+export default function InstanceSizesPage() {
     const { t } = useTranslation(['admin', 'common']);
-    const [page, setPage] = useState(1);
     const [globalSearch, setGlobalSearch] = useState('');
     const deferredSearch = useDeferredValue(globalSearch);
     const isStale = globalSearch !== deferredSearch;
@@ -85,41 +98,29 @@ export default function TemplatesPage() {
     const [searchedColumn, setSearchedColumn] = useState('');
     const [searchText, setSearchText] = useState('');
 
-    // Fetch templates
-    const { data, isLoading, refetch } = useApiGet<TemplateList>(
-        ['admin-templates', page],
-        () => api.GET('/templates', {
-            params: { query: { page, per_page: 20 } },
-        })
+    // Fetch instance sizes
+    const { data, isLoading, refetch } = useApiGet<InstanceSizeList>(
+        ['admin-instance-sizes'],
+        () => api.GET('/instance-sizes')
     );
-
-    // Derive unique OS families for column filters
-    const osFamilyFilters = useMemo(() => {
-        const families = new Set<string>();
-        (data?.items ?? []).forEach((tpl) => {
-            if (tpl.os_family) families.add(tpl.os_family);
-        });
-        return Array.from(families).sort().map((f) => ({ text: f, value: f }));
-    }, [data?.items]);
 
     // useMemo: cache filtered results based on deferred search value
     const filteredItems = useMemo(() => {
         const items = data?.items ?? [];
         if (!deferredSearch) return items;
         const q = deferredSearch.toLowerCase();
-        return items.filter((tpl) =>
-            tpl.name.toLowerCase().includes(q) ||
-            (tpl.display_name ?? '').toLowerCase().includes(q) ||
-            (tpl.description ?? '').toLowerCase().includes(q) ||
-            (tpl.os_family ?? '').toLowerCase().includes(q)
+        return items.filter((sz) =>
+            sz.name.toLowerCase().includes(q) ||
+            (sz.display_name ?? '').toLowerCase().includes(q) ||
+            (sz.description ?? '').toLowerCase().includes(q) ||
+            getCapabilityLabels(sz).some((c) => c.toLowerCase().includes(q))
         );
     }, [data?.items, deferredSearch]);
 
     /**
      * Ant Design best practice: filterDropdown for free-text column search.
-     * Supports Enter to confirm + Reset to clear.
      */
-    const getColumnSearchProps = (dataIndex: keyof Template): Partial<ColumnsType<Template>[number]> => ({
+    const getColumnSearchProps = (dataIndex: keyof InstanceSize): Partial<ColumnsType<InstanceSize>[number]> => ({
         filterDropdown: ({ setSelectedKeys, selectedKeys, confirm, clearFilters }: FilterDropdownProps) => (
             <div style={{ padding: 8 }} onKeyDown={(e) => e.stopPropagation()}>
                 <Input
@@ -168,23 +169,25 @@ export default function TemplatesPage() {
         ),
         onFilter: (value, record) =>
             (record[dataIndex] ?? '').toString().toLowerCase().includes((value as string).toLowerCase()),
-        onFilterDropdownOpenChange: (visible) => {
-            if (visible) {
-                setTimeout(() => searchInputRef.current?.select(), 100);
-            }
+        filterDropdownProps: {
+            onOpenChange: (visible) => {
+                if (visible) {
+                    setTimeout(() => searchInputRef.current?.select(), 100);
+                }
+            },
         },
     });
 
-    const columns: ColumnsType<Template> = [
+    const columns: ColumnsType<InstanceSize> = [
         {
             title: t('common:table.name'),
             dataIndex: 'name',
             key: 'name',
             ...getColumnSearchProps('name'),
             sorter: (a, b) => a.name.localeCompare(b.name),
-            render: (name: string, record: Template) => (
+            render: (name: string, record: InstanceSize) => (
                 <Space>
-                    <FileTextOutlined style={{ color: '#1677ff' }} />
+                    <HddOutlined style={{ color: '#1677ff' }} />
                     <div>
                         <Text strong>
                             {searchedColumn === 'name'
@@ -200,39 +203,89 @@ export default function TemplatesPage() {
             ),
         },
         {
-            title: t('templates.os_family'),
-            dataIndex: 'os_family',
-            key: 'os_family',
-            width: 120,
-            // Ant Design best practice: enum filters with onFilter
-            filters: osFamilyFilters,
-            onFilter: (value, record) => record.os_family === value,
-            render: (family: string | undefined) => {
-                if (!family) return <Text type="secondary">—</Text>;
-                const color = OS_COLOR_MAP[family.toLowerCase()] ?? 'default';
-                return <Tag color={color}>{family}</Tag>;
-            },
-        },
-        {
-            title: t('templates.os_version'),
-            dataIndex: 'os_version',
-            key: 'os_version',
-            width: 120,
-            render: (v: string | undefined) => v ? <Tag>{v}</Tag> : '—',
-        },
-        {
-            title: t('templates.version'),
-            dataIndex: 'version',
-            key: 'version',
-            width: 90,
+            title: t('instanceSizes.cpu'),
+            dataIndex: 'cpu_cores',
+            key: 'cpu_cores',
+            width: 100,
             align: 'center' as const,
-            sorter: (a, b) => a.version - b.version,
-            render: (v: number) => (
-                <Tag color="processing">v{v}</Tag>
+            sorter: (a, b) => a.cpu_cores - b.cpu_cores,
+            render: (cores: number, record: InstanceSize) => (
+                <Space direction="vertical" size={0} style={{ textAlign: 'center' }}>
+                    <Text strong>{cores} {t('instanceSizes.cores')}</Text>
+                    {record.dedicated_cpu && (
+                        <Tag color="orange" style={{ fontSize: 10 }}>
+                            <ThunderboltOutlined /> {t('instanceSizes.dedicated')}
+                        </Tag>
+                    )}
+                </Space>
             ),
         },
         {
-            title: t('templates.enabled'),
+            title: t('instanceSizes.memory'),
+            dataIndex: 'memory_mb',
+            key: 'memory_mb',
+            width: 100,
+            align: 'center' as const,
+            sorter: (a, b) => a.memory_mb - b.memory_mb,
+            render: (mb: number) => <Text strong>{formatMemory(mb)}</Text>,
+        },
+        {
+            title: t('instanceSizes.disk'),
+            dataIndex: 'disk_gb',
+            key: 'disk_gb',
+            width: 100,
+            align: 'center' as const,
+            sorter: (a, b) => (a.disk_gb ?? 0) - (b.disk_gb ?? 0),
+            render: (gb: number | undefined) => gb ? <Text>{gb} GB</Text> : <Text type="secondary">—</Text>,
+        },
+        {
+            title: t('instanceSizes.capabilities'),
+            key: 'capabilities',
+            width: 220,
+            // Ant Design best practice: multi-select enum column filter
+            filters: [
+                { text: 'GPU', value: 'gpu' },
+                { text: 'SR-IOV', value: 'sriov' },
+                { text: 'Hugepages', value: 'hugepages' },
+                { text: 'Dedicated CPU', value: 'dedicated' },
+            ],
+            onFilter: (value, record) => {
+                switch (value) {
+                    case 'gpu': return !!record.requires_gpu;
+                    case 'sriov': return !!record.requires_sriov;
+                    case 'hugepages': return !!record.requires_hugepages;
+                    case 'dedicated': return !!record.dedicated_cpu;
+                    default: return false;
+                }
+            },
+            filterMultiple: true,
+            render: (_: unknown, record: InstanceSize) => {
+                const tags: React.ReactNode[] = [];
+                if (record.requires_gpu) {
+                    tags.push(<Tag key="gpu" color="volcano">GPU</Tag>);
+                }
+                if (record.requires_sriov) {
+                    tags.push(<Tag key="sriov" color="purple">SR-IOV</Tag>);
+                }
+                if (record.requires_hugepages) {
+                    tags.push(
+                        <Tag key="hugepages" color="cyan">
+                            Hugepages {record.hugepages_size ? `(${record.hugepages_size})` : ''}
+                        </Tag>
+                    );
+                }
+                if (record.dedicated_cpu) {
+                    tags.push(
+                        <Tag key="dedicated" color="orange">
+                            <ThunderboltOutlined /> Dedicated
+                        </Tag>
+                    );
+                }
+                return tags.length > 0 ? <Space wrap>{tags}</Space> : <Text type="secondary">—</Text>;
+            },
+        },
+        {
+            title: t('instanceSizes.enabled'),
             dataIndex: 'enabled',
             key: 'enabled',
             width: 90,
@@ -268,8 +321,8 @@ export default function TemplatesPage() {
                 marginBottom: 24,
             }}>
                 <div>
-                    <Title level={4} style={{ margin: 0 }}>{t('templates.title')}</Title>
-                    <Text type="secondary">{t('templates.subtitle')}</Text>
+                    <Title level={4} style={{ margin: 0 }}>{t('instanceSizes.title')}</Title>
+                    <Text type="secondary">{t('instanceSizes.subtitle')}</Text>
                 </div>
                 <Space>
                     <Input
@@ -292,27 +345,25 @@ export default function TemplatesPage() {
                     opacity: isStale ? 0.6 : 1,
                     transition: isStale ? 'opacity 0.2s 0.1s linear' : 'opacity 0s 0s linear',
                 }}>
-                    <Table<Template>
+                    <Table<InstanceSize>
                         columns={columns}
                         dataSource={filteredItems}
                         rowKey="id"
                         loading={isLoading}
+                        size="middle"
                         pagination={{
-                            current: page,
-                            total: data?.pagination?.total ?? filteredItems.length,
+                            total: filteredItems.length,
                             pageSize: 20,
-                            onChange: setPage,
                             showTotal: (total) => t('common:table.total', { total }),
                             showSizeChanger: false,
                         }}
-                        size="middle"
                         locale={{
                             emptyText: (
                                 <Empty
                                     description={
                                         deferredSearch
                                             ? t('common:message.no_data')
-                                            : t('templates.empty')
+                                            : t('instanceSizes.empty')
                                     }
                                     image={Empty.PRESENTED_IMAGE_SIMPLE}
                                 />
