+++ web/src/app/(protected)/admin/namespaces/page.tsx
