+++ web/src/app/(protected)/dashboard/layout.tsx
@@ -1,8 +1,7 @@
 import type { Metadata } from 'next';
-import ProtectedLayout from '@/components/layouts/ProtectedLayout';
 
 export const metadata: Metadata = {
-    title: 'Dashboard - KubeVirt Shepherd',
+    title: 'Dashboard',
 };
 
 export default function DashboardLayout({
@@ -10,5 +9,5 @@ export default function DashboardLayout({
 }: {
     children: React.ReactNode;
 }) {
-    return <ProtectedLayout>{children}</ProtectedLayout>;
+    return <>{children}</>;
 }
